requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==10.4.0
ultralytics==8.3.8
pytest==8.3.3
requests==2.32.3
httpx
